+++ R/spatial.R
@@ -1,25 +1,28 @@
 diag_expand <- function(...) UseMethod("diag_expand")
 
-diag_expand.default <- function(m, times=NULL, is.array=FALSE, auto.lag=TRUE) {
+diag_expand.default <- function(m, times, auto.lag=TRUE) {
+
+  # Checking class
+  meta <- classify_graph(m)
 
   # Getting the info
-  d <- dim(m)
+  d <- with(meta, c(n, n, nper))
+  if (missing(times)) times <- d[3]
 
-  if (is.array) times <- d[3]
+  if (!times)
+    stop("It must be a dynamic graph. nslices() = 0.")
 
   # Structure
-  W <- methods::as(Matrix::sparseMatrix(
-    i={}, j={},
-    dims=d[1:2]*times,
-    giveCsparse = TRUE), "dgCMatrix")
+  W <- methods::new("dgCMatrix", Dim=d[1:2]*times, p=rep(0L,d[1]*times+1L))
 
   # Filling
   for (p in 1:times) {
     i <- ((p-1)*d[1]+1):(d[1]*p)
     j <- ((p-1)*d[2]+1):(d[2]*p)
 
-    if (is.array) W[i,j] <- m[,,p]
-    else W[i,j] <- m
+    if (meta$class=="array") W[i,j] <- m[,,p]
+    else if (meta$class=="list") W[i,j] <- m[[p]]
+    else if (meta$class=="matrix") W[i,j] <- m
   }
 
   # Autolag
@@ -29,22 +32,5 @@ diag_expand.default <- function(m, times=NULL, is.array=FALSE, auto.lag=TRUE) {
 }
 
 diag_expand.diffnet <- function(g, ...) {
-  d     <- rep(nnodes(g),2)
-  times <- nslices(g)
-
-  # Casket
-  W <- methods::as(Matrix::sparseMatrix(i={}, j={}, dims = d*times),
-                   "dgCMatrix")
-
-  # Filling
-  for (p in 1:times) {
-    i <- ((p-1)*d[1]+1):(d[1]*p)
-    j <- ((p-1)*d[2]+1):(d[2]*p)
-    W[i,j] <- g$graph[[p]]
-  }
-
-  # Autolag
-  al <- cbind(1:d[1], 1:d[2])
-
-  W
+  diag_expand.default(g$graph, g$meta$nper)
 }