+++ src/stats.h
@@ -7,15 +7,12 @@
 #define NETDIFFUSER_STATS_
 using namespace Rcpp;
 
-arma::mat cumulative_adopt_count_cpp(const arma::mat & cumadopt);
-
-arma::rowvec hazard_rate_cpp(const arma::mat & cumadopt);
-
-arma::colvec threshold_cpp(const arma::mat & exposure,
-                           const arma::vec & toa, bool include_censored=false);
-
 arma::sp_mat vertex_covariate_dist(const arma::sp_mat & graph,
                                    const NumericMatrix & X, double p=2.0);
+
+arma::sp_mat vertex_mahalanobis_dist_cpp(const arma::sp_mat & graph,
+                                     const arma::mat & X,
+                                     const arma::mat & W);
 
 arma::sp_mat vertex_covariate_compare(const arma::sp_mat & graph, const NumericVector & X,
                                       std::string symbol);